requirements.txt
altgraph==0.17.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
macholib==1.16.3
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
plotly==5.24.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
setuptools==75.1.0
six==1.16.0
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.0.4
zipp==3.20.2
